# dose_prediction_system/openapi.reference.json
{"openapi":"3.1.0","info":{"title":"OpenKBP 头颈部放疗剂量预测系统","version":"0.1.0"},"paths":{"/":{"get":{"summary":"Index","operationId":"index__get","parameters":[{"name":"embedded","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Embedded"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/meta":{"get":{"summary":"Meta","operationId":"meta_api_meta_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/run-inference":{"post":{"summary":"Start Inference Job","operationId":"start_inference_job_api_run_inference_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_start_inference_job_api_run_inference_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/run-upload-inference":{"post":{"summary":"Start Upload Inference Job","operationId":"start_upload_inference_job_api_run_upload_inference_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_start_upload_inference_job_api_run_upload_inference_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/jobs/{job_id}":{"get":{"summary":"Get Job","operationId":"get_job_api_jobs__job_id__get","parameters":[{"name":"job_id","in":"path","required":true,"schema":{"type":"string","title":"Job Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/run-inference":{"post":{"summary":"Run Inference","operationId":"run_inference_run_inference_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_run_inference_run_inference_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/result/{result_name}":{"get":{"summary":"Result Page","operationId":"result_page_result__result_name__get","parameters":[{"name":"result_name","in":"path","required":true,"schema":{"type":"string","title":"Result Name"}},{"name":"embedded","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Embedded"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_run_inference_run_inference_post":{"properties":{"split":{"type":"string","title":"Split"},"patient_id":{"type":"string","title":"Patient Id"},"checkpoint_path":{"type":"string","title":"Checkpoint Path"}},"type":"object","required":["split","patient_id","checkpoint_path"],"title":"Body_run_inference_run_inference_post"},"Body_start_inference_job_api_run_inference_post":{"properties":{"split":{"type":"string","title":"Split"},"patient_id":{"type":"string","title":"Patient Id"},"checkpoint_path":{"type":"string","title":"Checkpoint Path"}},"type":"object","required":["split","patient_id","checkpoint_path"],"title":"Body_start_inference_job_api_run_inference_post"},"Body_start_upload_inference_job_api_run_upload_inference_post":{"properties":{"checkpoint_path":{"type":"string","title":"Checkpoint Path"},"case_zip":{"type":"string","format":"binary","title":"Case Zip"}},"type":"object","required":["checkpoint_path","case_zip"],"title":"Body_start_upload_inference_job_api_run_upload_inference_post"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}